requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
pymongo==3.11.3
python-alipay-sdk==3.0.1
waitress==2.0.0
